h26x/hevc: 12-bit + range-extension PPS fields, cu-lossless verified
SAO offsets scale by log2_sao_offset_scale (PPS range extension, 0 by
default) rather than the withdrawn bit-depth shift; that made Main 12
bit-exact. Parse pps_range_extension (transform-skip block size honoured;
cross-component prediction and chroma QP offset lists refused). Merge/AMVP
mvd greater1 uses the second context slot of the generated table. Reference
sets are DPB picture ids, not indices. Trace hooks read the environment once.

Nine more x265 fixtures pass bit-exact (open/closed GOP, slices+WPP+QG16+CIP,
lossless, cu-lossless, CTB16+temporal layers, odd size, no-TMVP/no-deblock,
Main 12). On the cu-lossless stream libavcodec 8.0 disagrees with x265's own
reconstruction and libde265 in the chroma of transquant-bypass CUs; this
decoder matches the encoder and libde265, so that fixture's reference is
the x265 --recon output.

+++ src/hevc/pic.rs
@@ -12,8 +12,8 @@ pub struct SaoParams {
     pub type_idx: u8,
     /// `sao_band_position` (band) or `SaoEoClass` (edge).
     pub band_or_class: u8,
-    /// `SaoOffsetVal[1..=4]` (already scaled by the bit-depth shift).
-    pub offsets: [i8; 4],
+    /// `SaoOffsetVal[1..=4]` (already shifted by `log2OffsetScale`).
+    pub offsets: [i16; 4],
 }
 
 /// Per-slice loop filter facts, indexed by the slice number stored per CTB.
